requirements.txt
numpy>=1.21.0
matplotlib>=3.4.3
networkx>=2.6.3
pandas>=1.3.0
seaborn>=0.11.2
tqdm>=4.62.3
openai>=1.0.0
python-dotenv>=0.19.0
